Major restore build!

+++ scripts/eastern_kingdoms/scarlet_enclave/ebon_hold.cpp
@@ -1196,7 +1196,7 @@ struct MANGOS_DLL_DECL npc_eye_of_acherusAI : public ScriptedAI
     {
         if(m_creature->isCharmed())
         {
-            if (StartTimer < uiDiff && !Active)
+            if (uint32(StartTimer) < uiDiff && !Active)
             {
                 m_creature->CastSpell(m_creature, 70889, true);
                 m_creature->CastSpell(m_creature, 51892, true);
@@ -1210,7 +1210,7 @@ struct MANGOS_DLL_DECL npc_eye_of_acherusAI : public ScriptedAI
                 StartTimer -= uiDiff;
         }
         else
-            if (StartTimer < uiDiff)
+            if (uint32(StartTimer) < uiDiff)
                 m_creature->ForcedDespawn();
     }
 };
